+++ site/src/components/BuildAvatar/BuildAvatar.tsx
